+++ models/insulated.msh
@@ -1,0 +1,214 @@
+$MeshFormat
+2.2 0 8
+$EndMeshFormat
+$PhysicalNames
+3
+1 3 "convection"
+2 1 "wire"
+2 2 "insulation"
+$EndPhysicalNames
+$Nodes
+67
+1 2 0 0
+2 -0.9999999999999996 1.732050807568877 0
+3 -1.000000000000001 -1.732050807568877 0
+4 3 0 0
+5 -1.499999999999999 2.598076211353316 0
+6 -1.500000000000001 -2.598076211353315 0
+7 1.827090914409903 0.8134732881175547 0
+8 1.338261209269959 1.486289654059163 0
+9 0.6180339846184854 1.902113033932683 0
+10 -0.2090569285631289 1.989043790523414 0
+11 -1.618033990162066 1.17557050264126 0
+12 -1.95629520245368 0.4158233769964312 0
+13 -1.956295200503979 -0.4158233861690545 0
+14 -1.618033987472603 -1.175570506342989 0
+15 -0.2090569245005841 -1.989043790950405 0
+16 0.6180339927037 -1.902113031305638 0
+17 1.33826121597086 -1.486289648025645 0
+18 1.827090916184073 -0.8134732841327028 0
+19 2.866718416712208 0.884265525327689 0
+20 2.478716320112386 1.689960178349925 0
+21 1.870469399717563 2.345494452076197 0
+22 1.096023066394168 2.792621248564138 0
+23 0.2241902761490784 2.991611391889027 0
+24 -0.6675628038033028 2.924783736103966 0
+25 -2.19915561691068 2.040518211781069 0
+26 -2.70290660585828 1.301651212886028 0
+27 -2.96649247971263 0.4471267916468595 0
+28 -2.966492477661765 -0.447126805253458 0
+29 -2.702906601876784 -1.301651221153692 0
+30 -2.199155614492735 -2.040518214386993 0
+31 -0.667562799591218 -2.924783737065347 0
+32 0.2241902856922546 -2.991611391173865 0
+33 1.096023079844227 -2.792621243285379 0
+34 1.870469411104438 -2.345494442995467 0
+35 2.478716325782583 -1.689960170033273 0
+36 2.866718418080356 -0.8842655208922591 0
+37 -0.9339546057726886 -0.4158233831706852 0
+38 0.02141496984582745 1.058980690221818 0
+39 -0.1941278312183717 0.1071477189577097 0
+40 0.9834805227413372 0.6826480240423473 0
+41 -0.7686480115959173 0.5453804900735302 0
+42 0.9999999999999992 -0.2125565610925706 0
+43 0.1213651863888595 -1.154712599348981 0
+44 -0.6406278644518437 -1.11916553237733 0
+45 0.8593162093855847 -0.9615150557898803 0
+46 -0.5943735880144831 1.234440025041052 0
+47 -1.335215615779825 0.1351111195641902 0
+48 0.6825711700739076 1.215293176376306 0
+49 1.45264285928781 0.3208911877668328 0
+50 0.3069952762974677 -0.3818091452905251 0
+51 -0.2680699677513154 -0.5928725882459622 0
+52 0.3727472045845057 0.474808656902285 0
+53 1.237382076390581 -2.143208614122736 0
+54 -2.474764143942451 -5.801802924119936e-09 0
+55 -0.7488578504306679 2.3047474739574 0
+56 1.237382067086159 2.143208619683771 0
+57 -1.621540936588582 -1.800903659450643 0
+58 -1.621540937918335 1.800903656784694 0
+59 2.37039878696343 -0.5038438145749483 0
+60 -2.187495154555313 -1.033152983087325 0
+61 0.1990108480843664 2.411002866027006 0
+62 1.988484306593702 -1.377849882960454 0
+63 -2.187495156981738 1.033152975378704 0
+64 1.988484301214432 1.377849888762947 0
+65 0.1990108553729037 -2.411002864348951 0
+66 -0.7488578478505082 -2.304747474025119 0
+67 2.370398785654296 0.5038438169983678 0
+$EndNodes
+$Elements
+132
+1 1 2 3 4 4 19
+2 1 2 3 4 19 20
+3 1 2 3 4 20 21
+4 1 2 3 4 21 22
+5 1 2 3 4 22 23
+6 1 2 3 4 23 24
+7 1 2 3 4 24 5
+8 1 2 3 5 5 25
+9 1 2 3 5 25 26
+10 1 2 3 5 26 27
+11 1 2 3 5 27 28
+12 1 2 3 5 28 29
+13 1 2 3 5 29 30
+14 1 2 3 5 30 6
+15 1 2 3 6 6 31
+16 1 2 3 6 31 32
+17 1 2 3 6 32 33
+18 1 2 3 6 33 34
+19 1 2 3 6 34 35
+20 1 2 3 6 35 36
+21 1 2 3 6 36 4
+22 2 2 1 1 39 38 41
+23 2 2 1 1 37 39 41
+24 2 2 1 1 38 39 52
+25 2 2 1 1 13 14 37
+26 2 2 1 1 9 10 38
+27 2 2 1 1 18 1 42
+28 2 2 1 1 15 16 43
+29 2 2 1 1 41 11 47
+30 2 2 1 1 11 41 46
+31 2 2 1 1 37 14 44
+32 2 2 1 1 18 42 45
+33 2 2 1 1 43 16 45
+34 2 2 1 1 15 43 44
+35 2 2 1 1 11 12 47
+36 2 2 1 1 39 37 51
+37 2 2 1 1 9 38 48
+38 2 2 1 1 42 1 49
+39 2 2 1 1 2 11 46
+40 2 2 1 1 38 10 46
+41 2 2 1 1 41 38 46
+42 2 2 1 1 13 37 47
+43 2 2 1 1 14 3 44
+44 2 2 1 1 17 18 45
+45 2 2 1 1 3 15 44
+46 2 2 1 1 16 17 45
+47 2 2 1 1 7 8 40
+48 2 2 1 1 37 41 47
+49 2 2 1 1 50 42 52
+50 2 2 1 1 42 40 52
+51 2 2 1 1 39 50 52
+52 2 2 1 1 10 2 46
+53 2 2 1 1 12 13 47
+54 2 2 1 1 40 8 48
+55 2 2 1 1 7 40 49
+56 2 2 1 1 40 42 49
+57 2 2 1 1 48 38 52
+58 2 2 1 1 40 48 52
+59 2 2 1 1 1 7 49
+60 2 2 1 1 8 9 48
+61 2 2 1 1 43 50 51
+62 2 2 1 1 50 39 51
+63 2 2 1 1 44 43 51
+64 2 2 1 1 43 45 50
+65 2 2 1 1 37 44 51
+66 2 2 1 1 45 42 50
+67 2 2 2 2 58 25 63
+68 2 2 2 2 36 59 62
+69 2 2 2 2 24 55 61
+70 2 2 2 2 67 19 64
+71 2 2 2 2 30 57 60
+72 2 2 2 2 66 31 65
+73 2 2 2 2 11 58 63
+74 2 2 2 2 7 67 64
+75 2 2 2 2 15 66 65
+76 2 2 2 2 55 10 61
+77 2 2 2 2 59 18 62
+78 2 2 2 2 57 14 60
+79 2 2 2 2 35 36 62
+80 2 2 2 2 23 24 61
+81 2 2 2 2 25 26 63
+82 2 2 2 2 29 30 60
+83 2 2 2 2 19 20 64
+84 2 2 2 2 31 32 65
+85 2 2 2 2 13 28 60
+86 2 2 2 2 9 22 61
+87 2 2 2 2 17 34 62
+88 2 2 2 2 21 8 64
+89 2 2 2 2 33 16 65
+90 2 2 2 2 27 12 63
+91 2 2 2 2 28 13 54
+92 2 2 2 2 22 9 56
+93 2 2 2 2 34 17 53
+94 2 2 2 2 8 21 56
+95 2 2 2 2 16 33 53
+96 2 2 2 2 12 27 54
+97 2 2 2 2 4 1 59
+98 2 2 2 2 5 2 55
+99 2 2 2 2 6 3 57
+100 2 2 2 2 1 4 67
+101 2 2 2 2 2 5 58
+102 2 2 2 2 3 6 66
+103 2 2 2 2 28 29 60
+104 2 2 2 2 20 21 64
+105 2 2 2 2 32 33 65
+106 2 2 2 2 34 35 62
+107 2 2 2 2 22 23 61
+108 2 2 2 2 26 27 63
+109 2 2 2 2 4 19 67
+110 2 2 2 2 5 25 58
+111 2 2 2 2 6 31 66
+112 2 2 2 2 30 6 57
+113 2 2 2 2 24 5 55
+114 2 2 2 2 36 4 59
+115 2 2 2 2 27 28 54
+116 2 2 2 2 21 22 56
+117 2 2 2 2 33 34 53
+118 2 2 2 2 13 12 54
+119 2 2 2 2 17 16 53
+120 2 2 2 2 9 8 56
+121 2 2 2 2 11 2 58
+122 2 2 2 2 7 1 67
+123 2 2 2 2 15 3 66
+124 2 2 2 2 3 14 57
+125 2 2 2 2 2 10 55
+126 2 2 2 2 1 18 59
+127 2 2 2 2 8 7 64
+128 2 2 2 2 12 11 63
+129 2 2 2 2 16 15 65
+130 2 2 2 2 18 17 62
+131 2 2 2 2 10 9 61
+132 2 2 2 2 14 13 60
+$EndElements
